fin quete 09

+++ suivi-des-drones-quete-008/suivi-des-drones.Core.Models/HealthStatus.cs
@@ -31,9 +31,10 @@ namespace suivi_des_drones.Core.Models
         public static HealthStatus Broken = new HealthStatus() { Id = -1, Label = "Cassé" };
         public static HealthStatus Repair = new HealthStatus() { Id = -2, Label = "En réparation" };
 
-
-        public int Id { get; init; }
+        #region Properties
+        public decimal Id { get; init; }
         public string Label { get; init; } = default!;
+        #endregion
     }
